[#3748] deleted <relatedProjects> & refactored
