+++ akvo/rsr/spa/app/utils/auto-save.jsx
@@ -70,6 +70,9 @@ class AutoSave extends React.Component {
       const item = thisValues
       const difference = customDiff(savedValues, item)
       delete difference.disaggregationTargets
+      if(setName === 'relatedProjects'){
+        if (Object.keys(item).indexOf('relatedProject') === -1 && Object.keys(item).indexOf('relatedIatiId') === -1) return
+      }
       // if difference is not empty AND the difference is not just the newly created item id inserted from ADDED_NEW_ITEM
       if (!isEmpty(difference)) {
         if(
